chore: move verify package next to eigenda v1
no reason for this to be top level now that we have v1 and v2 in same repo, and the verify package is only used for v1.
