+++ store/generated_key/eigenda/verify/certificate.go
