+++ store/generated_key/eigenda/verify/cli.go
